i forgot comit(

+++ backend/src/controllers/resource_controller.rs
@@ -1,19 +1,23 @@
 use rocket::{State, http::Status, serde::json::Json};
 use sea_orm::DatabaseConnection;
 
-use crate::{models::resource_model::{CreateResource, Model, ResourceDto, UpdateResource}, services::resources_service};
+use crate::{auth::guard::AuthUser, models::resource_model::{CreateResource, Model, ResourceDto, ResourceQuery, UpdateResource}, services::resources_service};
 
-#[get("/")]
-pub async fn get_all_resources(db: &State<DatabaseConnection>) -> Result<Json<Vec<ResourceDto>>, String>{
-    match resources_service::get_all(db).await {
+#[get("/?<query..>")]
+pub async fn get_all_resources(db: &State<DatabaseConnection>, query: ResourceQuery) -> Result<Json<Vec<ResourceDto>>, String>{
+    match resources_service::get_all(db, query).await {
         Ok(resources) => Ok(Json(resources)),
         Err(e) => Err(e.to_string())
     }
 }
 
 #[post("/", data="<dto>")]
-pub async fn create_resources(db: &State<DatabaseConnection>, dto: Json<CreateResource>) -> Result<Json<Model>, Status>{
-    match resources_service::create(db, dto.into_inner()).await {
+pub async fn create_resources(db: &State<DatabaseConnection>, dto: Json<CreateResource>, user: AuthUser) -> Result<Json<Model>, Status>{
+    if user.id == 0{
+        return Err(Status::NonAuthoritativeInformation);
+    }
+
+    match resources_service::create(db, dto.into_inner(), user).await {
         Ok(resource) => Ok(Json(resource)),
         Err(_) => Err(Status::InternalServerError)
     }
@@ -28,7 +32,11 @@ pub async fn get_one_resource(db: &State<DatabaseConnection>, id:i32) -> Result<
 }
 
 #[patch("/<id>", data="<update_dto>")]
-pub async fn update_resource(db: &State<DatabaseConnection>, id:i32, update_dto: Json<UpdateResource>) -> Result<Json<Model>, Status>{
+pub async fn update_resource(db: &State<DatabaseConnection>, id:i32, update_dto: Json<UpdateResource>, user: AuthUser) -> Result<Json<Model>, Status>{
+        if user.id == 0{
+        return Err(Status::NonAuthoritativeInformation);
+    }
+    
     match resources_service::update(db, id, update_dto.into_inner()).await {
         Ok(updated) => Ok(Json(updated)),
         Err(_) => Err(Status::NotFound)
@@ -36,7 +44,22 @@ pub async fn update_resource(db: &State<DatabaseConnection>, id:i32, update_dto:
 }
 
 #[delete("/<id>")]
-pub async fn delete_resource(db: &State<DatabaseConnection>, id:i32) -> Status{
+pub async fn delete_resource(db: &State<DatabaseConnection>, id:i32, user: AuthUser) -> Status{
+
+    if user.id == 0 {
+        return Status::NonAuthoritativeInformation;
+    }
+
+    let resource_user_id = match resources_service::get_one_model(db, id).await {
+        Ok(resource) => resource.user_id,
+        Err(_) => return Status::NotFound
+    };
+
+    if user.id != resource_user_id {
+        return Status::Forbidden;
+    }
+
+
     match resources_service::delete(db, id).await {
         Ok(_) => Status::NoContent,
         Err(_) => Status::NotFound
